+++ src/components/button/index.js
@@ -1,14 +1,14 @@
-export { default as BaseButton } from './basebutton/BaseButton';
-export { default as LoginButton } from './loginbutton/LoginButton';
-export { DashboardButton } from './dashboardbutton/DashboardButton';
-export { DashboardAddButton } from './dashboardaddbutton/DashboardAddButton';
-export { DashboardDeleteButton } from './dashboarddeletebutton/DashboardDeleteButton';
-export { ModalCancelButton, ModalConfirmButton } from './modalbutton/ModalButton';
-export { AddColumnButton } from './addcolumnbutton/AddColumnButton';
-export { AddTodoButton } from './addtodobutton/AddTodoButton';
-export { InviteAcceptButton } from './inviteacceptbutton/InviteAcceptButton';
-export { InviteRejectButton } from './inviterejectbutton/InviteRejectButton';
-export { DeleteButton } from './deletebutton/DeleteButton';
-export { InputSubmitButton } from './inputsubmitbutton/InputSubmitButton';
-export { PaginationButton } from './paginationbutton/PaginationButton';
-export { default as PaginationPairButton } from './paginationpairbutton/PaginationPairButton';
+export { default as BaseButton } from './BaseButton/BaseButton';
+export { default as LoginButton } from './LoginButton/LoginButton';
+export { DashboardButton } from './DashboardButton/DashboardButton';
+export { DashboardAddButton } from './DashboardAddButton/DashboardAddButton';
+export { DashboardDeleteButton } from './DashboardDeleteButton/DashboardDeleteButton';
+export { ModalCancelButton, ModalConfirmButton } from './ModalButton/ModalButton';
+export { AddColumnButton } from './AddColumnButton/AddColumnButton';
+export { AddTodoButton } from './AddTodoButton/AddTodoButton';
+export { InviteAcceptButton } from './InviteAcceptButton/InviteAcceptButton';
+export { InviteRejectButton } from './InviteRejectButton/InviteRejectButton';
+export { DeleteButton } from './DeleteButton/DeleteButton';
+export { InputSubmitButton } from './InputsubmitButton/InputSubmitButton';
+export { PaginationButton } from './PaginationButton/PaginationButton';
+export { default as PaginationPairButton } from './PaginationpairButton/PaginationPairButton';
