+++ src/pages/mypage/index.js
@@ -3,7 +3,7 @@ import { useEffect } from 'react';
 import { useHeader } from '@/context/HeaderProvider';
 import ProfileCard from '@/components/cardtable/ProfileCard';
 import ChangePasswordCard from '@/components/cardtable/ChangePasswordCard';
-import BackPageButton from '@/components/button/backpagebutton/BackPageButton';
+import BackPageButton from '@/components/button/BackPageButton/BackPageButton';
 
 export default function MyPage() {
   const { setHeaderConfig } = useHeader();
